v.16.9.2
add check if there is parameters that give negative unconditional
variance. There was a  lot of warning during bayesian estimation because
this check was not in place (the square root of a negative value gave a
warning). Now when there is negative unconditional variance, the value
is replaced by an NA before the warning is triggered. In a next
extension where the model will allow for explosive state, this will be
change.